mark some important types readonly
instances of these types are usually shared across multiple independent runs and it would lead to very unexpected results if they were to change, so enforce that they are immutable.

note that courses do have to be mutated when loaded, to sort slopes. provide a getCourse() helper function to do this and return an immutable CourseData object.