add some additional no-op random conditions
the problem with making these random is that, when combined with a dynamic condition, it can wrongly cause the dynamic condition to not activate. for example, take the case of 先頭プライド, running_style==1&phase<=1&change_order_onetime>0&accumulatetime>=5@running_style==1&phase<=1&blocked_side_continuetime>=2&accumulatetime>=5. with this implementation, it statically picks a random trigger and then checks if the dynamic condition (accumulatetime) is satisfied there. this means if the trigger is before accumulatetime is satisifed, the skill with not activate, when in the game change_order_onetime and blocked_side_continuetime are what we would call dynamic conditions, in which case the skill simply activates when they are all (dynamically) satisfied, and so the skill cannot fail to activate in this way.

in practice, this is unlikely to matter much, i think.