implement a new random sample policy with a log-normal distribution
the idea here is that some non-random conditions nevertheless behave kind of randomly, but tend to be biased towards activating at the beginning of their range. it is not at all clear that a log-normal distribution is the correct model for this.

currently this is used for conditions that involve other proximity to other umas, such as behind_near_lane_time, blocked_side_continuetime, change_order_onetime, and infront_near_lane_time. (notably this also implements behind_near_lane_time.)

it is worth noting that this more or less completely breaks skills that combine these conditions with accumulatetime, which is a lot of them actually, for example 先頭プライド、遊びはおしまいっ！、ウマ好み、スリップストリーム, since they were already somewhat broken but it was masked by the fact that only a few samples tended to be too early (within the accumulatetime window), but now many of them are. it's not at all clear to me how to resolve this with the current architecture.

LogNormalRandomPolicy is dominated by normal RandomPolicy because the latter is more stringent, i.e. that is how those skills actually work, instead of being a model for a not-actually-random process like the log-normal ones, and it would not make sense for (e.g.) phase_random to have a non-uniform distribution when combined with anything. regardless, i don't think there are any skills currently that combine _random with any of the conditions currently modeled as log-normal. in the future if there are any other conditions that we want to model as uniform random but not dominate log-normal conditions it might make sense to split RandomPolicy into the actual policy for _random ones and an identical UniformRandomPolicy for modeling non-random conditions that are modeled as uniformly random, with the only difference being how they reconcile with other policies. compete_fight_count comes to mind as a condition where that might make sense.