implement sample policy based on an Erlang distribution
replace the log-normal conditions with this, since it makes a bit more sense to me and seems to result in numbers closer to hoffe's. keep the actual log-normal code around though to allow the distribution to be configurable in the future.

this also does a bit of refactoring to make it easier to implement sample policies based on other distributions in the future.

the k=3 λ=2.0 parameters to the distribution used for infront_near_lane_time etc were determined experimentally by trying to find something resulting in close to hoffe's numbers for nonstop girl.