fix a problem where combining randomly-modeled conditions with accumulatetime would result in a substantial number of incorrectly failed activations
formerly it would place a random 10m trigger based on the underlying model distribution, and this trigger would often fall within the accumulatetime window, causing the skill to fail to activate even though this does not accurately reflect how it works in the game. however, there's no reason the trigger has to be 10m. fix this by simply extending the trigger to the end of the region it is placed in, so it will activate as soon as the accumulatetime condition is fulfilled.

technically, this is probably still not completely correct, since it skews activations to be immediately after the necessary accumulatetime instead of reflecting the assumed distribution. however, this is probably fine in practice, since the relevant skills usually activate quite early anyway.