add recoveries as a skill type (no-op)
currently does nothing. in preparation for implementing activate_count_heal.

the display in plot.py is kind of wrong since heals never deactivate so the bar just extends to the remainder of the graph. i don't just want to make it a fixed width since some heals do have other effects and therefore deactivate and i think it would be confusing if the display showed actual duration in those cases and an arbitrary width in others. the alternative is making the block really short which might look ugly. possibly i'll just end up drawing it as a vertical line or something.