implement pace down
position keep mechanic for non-runners. simulates a runner simultaneously to the main uma and uses that to check pace down thresholds. this is mostly intended for the beginning of the race where backlines starting in pace down is almost guaranteed, and affects the strength of skills such as ウマ好み.

pace down is only simulated for the first 5 sections (5/24 of the course) instead of 10 which is what is actually used in the game. past the beginning of the race pace down is too dependent on other umas and imo doesn't make sense to simulate.

NB. currently there's a caveat that since determining pace down exit uses Math.random() it won't be synchronized between the RaceSolver instances in gain.ts, which ends up producing some variation in the numbers. this will be fixed shortly.