have RaceSolver accept a PRNG for any random operations it needs to do
this allows deterministic comparisons between solver instances, which is basically what this whole thing is for anyway.

currently these are just used for checking the end of pace down and for rolling a random gold skill for summer goldship unique. e.g. int effects in the future will probably use an additional separate rng, to allow for changing int but keeping other random things deterministic.

fixes the caveat mentioned in the commit that introduced pace down.