use a separate prng for choosing random gold skills for 564
previously the presence/absence of 564 would have an affect on pace down exit rng. the principle is that everything that can be changed independently should have separate rng instances to maintain as much determinism as possible.

there are probably various merits and demerits to the RaceSolver constructing its own rng instances from the one provided instead of the caller providing all of them. however, i think this is fine since as long as two RaceSolvers are provided with rngs seeded with the same seed their internal state will be the same.