fix a bug where an uphill during start dash would cause speed to plummet
this actually does a bit more than that and changes it so that your target speed during start dash is no longer fixed at 0.85 * BaseSpeed. it turns out that if there is an uphill during start dash your target speed can be lower than the 0.85 * BaseSpeed threshold to break out of start dash. experimentally this seems to be what actually happens in the game and you indeed don't leave start dash until your target speed is high enough to get over the threshold.