add an approximation of the new 脚を溜める mechanic
this cheats quite a lot and just treats it as a 3s accel skill with a modifier scaling based on *displayed* power. the actual mechanic is more complicated with duration scaling, but how it actually works is currently unknown (?).

the modifier formula is also partially approximate, and obtained based on packet capture data.