estimate hp consumption for the hp_per condition
we're still not actually simulating hp. rather, we just take the lower bound of their activation region to be where hp would dip below hp_per if running at the course base speed. this is of course wildly inaccurate, especially for low hp_per values.

this is pretty much for one purpose and one purpose only, which is to keep 火事場のバ鹿力 from activating out of the gate and getting a huge bashin gain.