manually implement scaling for scenario skills
in general, with the exception of maybe the aoharu ones, these should all be fulfilled, so it's more accurate to assume their max scaling values.

note that we don't implement scaling at all currently, so just manually patch their modifiers. when we do eventually implement scaling we'll just have to always assume max values for these skills since their scaling conditions are external to the uma or the race.