propagate perspective information through to RaceSolver
this is so that it can be passed to the onSkillActivate/onSkillDeactivate callbacks, and we need access to it there to display skill activation information associated with the correct uma in the umalator (i.e. perspective == Other should display differently in comparisons from perspective == Self; this is the most general way i could think of to do this that encompasses buffs, debuffs, skills with both regular and debuff components, and skills with a self debuff)

perspective field is optional because it isn't actually used in RaceSolver, only passed to callbacks. the legacy (i.e. non-umalator) tools don't really track perspective information and don't use it anyway