implement random_lot and post_number as dynamic conditions
after thinking for a while there's really no good way to implement these statically, as the samplePolicy for a condition can't depend on the filter parameters. i considered changing that to be possible, but then some work might have to be done for lucky seven to properly use the joint distribution of a hypothetical RandomLotSamplePolicy and a PostNumberSamplePolicy, and in general samplePolicies dependent on filter parameters introduce some new headaches with reconciliation in the general case i think. possibly some work could be done in the future to reconcile sample policies better (can we just sample from both and take triggers where they overlap?) but for the time being it seems cleanest to implement these as dynamic conditions, and avoids any rng desync issues.